Move chi-squared helpers back into engine
